Remove RABBIT_ECHO
This is a "pit of doom".  It is common for callers to issue commands that interact with jq.
jq wants to read everything from stdin.  The messages we were emitting to aide in debugging would attempt to be parsed as json.  I tried several different approaches (e.g. using stderr for the echoing), but all attempts yielded unacceptable results.